changed read out of transpose block and...
added brams for transpose instead of the ibob srams.
if decide to increase the number of channels, may need to consider
other options.
read out is now all antennas for a single channel and all timesamples.
but spits out 128 channels for a single xengine first all antennas/all
time samples. then all 128 channels for second xengine ll antennas/all time.
so on and so forth.