During an unexpected matlab crash roachf_2048ch.mdl was
deleted. Luckily there was a backup file for the deleted modle file.
this file was renamed to roachf_2048_bak.mdl (included in this commit).
continuing from the previous commit to roachf_2048ch.mdl this mdl now contains
a cnts block for the 10gbe switch receive packets and an update to the 10gberxsnap block
in switch block that stores valids.