added new roachfengine model..
This model is an updated version of roachf_2048ch_bak.mdl
This model actually only has 1024 channels, and has a mux
for the number of xengines we have on the other side of the switch,
creating new packet formats on the fly. No need to recompile for increasing
the number of total inputs in the correlator. Just pick a different output of the mux.